AlertDialogsFragment: move process() method to AlertDialogsViewModel
